+++ hydra/src/process_registry.rs
@@ -13,9 +13,9 @@ use crate::SystemMessage;
 /// The global process registry.
 pub struct ProcessRegistry {
     /// A collection of process id -> process registration.
-    pub processes: HashMap<u32, ProcessRegistration>,
+    pub processes: HashMap<u64, ProcessRegistration>,
     /// A collection of registered named processes.
-    pub named_processes: HashMap<String, u32>,
+    pub named_processes: HashMap<String, u64>,
 }
 
 impl ProcessRegistry {
